♻️ refactor: move working sidebar into features
