+++ src/features/Conversation/WorkingSidebar/ResourcesSection/AgentDocumentsGroup.tsx
