+++ src/features/Conversation/WorkingSidebar/ResourcesSection/index.tsx
